+++ src/GSS.Authorization.OAuth.HttpClient/ServiceCollectionExtensions.cs
@@ -14,13 +14,15 @@ namespace GSS.Authorization.OAuth
         /// </summary>
         /// <typeparam name="TClient">The type of the typed client.</typeparam>
         /// <typeparam name="TRequestSigner">The type of the request signer.</typeparam>
+        /// <typeparam name="TOptions">The type of the configure options.</typeparam>
         /// <param name="services">The <see cref="IServiceCollection"/>.</param>
         /// <param name="configureOptions">A delegate that is used to configure an <see cref="OAuthHttpHandlerOptions"/>.</param>
         /// <returns>An <see cref="T:Microsoft.Extensions.DependencyInjection.IHttpClientBuilder" /> that can be used to configure the client.</returns>
-        public static IHttpClientBuilder AddOAuthHttpClient<TClient, TRequestSigner>(this IServiceCollection services,
-            Action<IServiceProvider, OAuthHttpHandlerOptions> configureOptions)
+        public static IHttpClientBuilder AddOAuthHttpClient<TClient, TRequestSigner, TOptions>(this IServiceCollection services,
+            Action<IServiceProvider, TOptions> configureOptions)
             where TClient : class
             where TRequestSigner : RequestSignerBase
+            where TOptions : OAuthHttpHandlerOptions
         {
             if (services == null)
             {
@@ -34,7 +36,7 @@ namespace GSS.Authorization.OAuth
 
             services.TryAddSingleton<TRequestSigner>();
 
-            services.AddOptions<OAuthHttpHandlerOptions>()
+            services.AddOptions<TOptions>()
                 .Configure<IServiceProvider>((options, resolver) => configureOptions(resolver, options))
                 .PostConfigure(ValidateOptions);
 
@@ -43,6 +45,38 @@ namespace GSS.Authorization.OAuth
                 .AddHttpMessageHandler(resolver => new OAuthHttpHandler(
                     resolver.GetRequiredService<IOptions<OAuthHttpHandlerOptions>>(),
                     resolver.GetRequiredService<TRequestSigner>()));
+        }
+
+        /// <summary>
+        /// Add typed HttpClient with <see cref="OAuthHttpHandler"/> and related services
+        /// </summary>
+        /// <typeparam name="TClient">The type of the typed client.</typeparam>
+        /// <typeparam name="TRequestSigner">The type of the request signer.</typeparam>
+        /// <param name="services">The <see cref="IServiceCollection"/>.</param>
+        /// <param name="configureOptions">A delegate that is used to configure an <see cref="OAuthHttpHandlerOptions"/>.</param>
+        /// <returns>An <see cref="T:Microsoft.Extensions.DependencyInjection.IHttpClientBuilder" /> that can be used to configure the client.</returns>
+        public static IHttpClientBuilder AddOAuthHttpClient<TClient, TRequestSigner>(this IServiceCollection services,
+            Action<IServiceProvider, OAuthHttpHandlerOptions> configureOptions)
+            where TClient : class
+            where TRequestSigner : RequestSignerBase
+        {
+            return services.AddOAuthHttpClient<TClient, TRequestSigner, OAuthHttpHandlerOptions>(configureOptions);
+        }
+
+        /// <summary>
+        /// Add typed HttpClient with <see cref="OAuthHttpHandler"/> and default request signer (HMAC-SHA1)
+        /// </summary>
+        /// <typeparam name="TClient">The type of the typed client.</typeparam>
+        /// <typeparam name="TOptions">The type of the configure options.</typeparam>
+        /// <param name="services">The <see cref="IServiceCollection"/>.</param>
+        /// <param name="configureOptions">A delegate that is used to configure an <see cref="OAuthHttpHandlerOptions"/>.</param>
+        /// <returns>An <see cref="T:Microsoft.Extensions.DependencyInjection.IHttpClientBuilder" /> that can be used to configure the client.</returns>
+        public static IHttpClientBuilder AddOAuthHttpClient<TClient, TOptions>(this IServiceCollection services,
+            Action<IServiceProvider, TOptions> configureOptions)
+            where TClient : class
+            where TOptions : OAuthHttpHandlerOptions
+        {
+            return services.AddOAuthHttpClient<TClient, HmacSha1RequestSigner, TOptions>(configureOptions);
         }
 
         /// <summary>
@@ -56,7 +90,7 @@ namespace GSS.Authorization.OAuth
             Action<IServiceProvider, OAuthHttpHandlerOptions> configureOptions)
             where TClient : class
         {
-            return services.AddOAuthHttpClient<TClient, HmacSha1RequestSigner>(configureOptions);
+            return services.AddOAuthHttpClient<TClient, OAuthHttpHandlerOptions>(configureOptions);
         }
 
         /// <summary>
@@ -65,11 +99,13 @@ namespace GSS.Authorization.OAuth
         /// <param name="services">The <see cref="IServiceCollection"/>.</param>
         /// <param name="name">The logical name of the <see cref="HttpClient"/> to configure.</param>
         /// <typeparam name="TRequestSigner">The type of the request signer.</typeparam>
+        /// <typeparam name="TOptions">The type of the configure options.</typeparam>
         /// <param name="configureOptions">A delegate that is used to configure an <see cref="OAuthHttpHandlerOptions"/>.</param>
         /// <returns>An <see cref="T:Microsoft.Extensions.DependencyInjection.IHttpClientBuilder" /> that can be used to configure the client.</returns>
-        public static IHttpClientBuilder AddOAuthHttpClient<TRequestSigner>(this IServiceCollection services, string name,
-            Action<IServiceProvider, OAuthHttpHandlerOptions> configureOptions)
+        public static IHttpClientBuilder AddOAuthHttpClient<TRequestSigner, TOptions>(this IServiceCollection services, string name,
+            Action<IServiceProvider, TOptions> configureOptions)
             where TRequestSigner : RequestSignerBase
+            where TOptions : OAuthHttpHandlerOptions
         {
             if (services == null)
             {
@@ -83,7 +119,7 @@ namespace GSS.Authorization.OAuth
 
             services.TryAddSingleton<TRequestSigner>();
 
-            services.AddOptions<OAuthHttpHandlerOptions>()
+            services.AddOptions<TOptions>()
                 .Configure<IServiceProvider>((options, resolver) => configureOptions(resolver, options))
                 .PostConfigure(ValidateOptions);
 
@@ -92,6 +128,36 @@ namespace GSS.Authorization.OAuth
                 .AddHttpMessageHandler(resolver => new OAuthHttpHandler(
                     resolver.GetRequiredService<IOptions<OAuthHttpHandlerOptions>>(),
                     resolver.GetRequiredService<TRequestSigner>()));
+        }
+
+        /// <summary>
+        /// Add named HttpClient with <see cref="OAuthHttpHandler"/> and related services
+        /// </summary>
+        /// <param name="services">The <see cref="IServiceCollection"/>.</param>
+        /// <param name="name">The logical name of the <see cref="HttpClient"/> to configure.</param>
+        /// <typeparam name="TRequestSigner">The type of the request signer.</typeparam>
+        /// <param name="configureOptions">A delegate that is used to configure an <see cref="OAuthHttpHandlerOptions"/>.</param>
+        /// <returns>An <see cref="T:Microsoft.Extensions.DependencyInjection.IHttpClientBuilder" /> that can be used to configure the client.</returns>
+        public static IHttpClientBuilder AddOAuthHttpClient<TRequestSigner>(this IServiceCollection services,
+            string name,
+            Action<IServiceProvider, OAuthHttpHandlerOptions> configureOptions)
+            where TRequestSigner : RequestSignerBase
+        {
+            return services.AddOAuthHttpClient<TRequestSigner, OAuthHttpHandlerOptions>(name, configureOptions);
+        }
+
+        /// <summary>
+        /// Add named HttpClient with <see cref="OAuthHttpHandler"/> and default request signer (HMAC-SHA1)
+        /// </summary>
+        /// <param name="services">The <see cref="IServiceCollection"/>.</param>
+        /// <param name="name">The logical name of the <see cref="HttpClient"/> to configure.</param>
+        /// <param name="configureOptions">A delegate that is used to configure an <see cref="OAuthHttpHandlerOptions"/>.</param>
+        /// <returns>An <see cref="T:Microsoft.Extensions.DependencyInjection.IHttpClientBuilder" /> that can be used to configure the client.</returns>
+        public static IHttpClientBuilder AddOAuthHttpClient<TOptions>(this IServiceCollection services, string name,
+            Action<IServiceProvider, TOptions> configureOptions)
+            where TOptions : OAuthHttpHandlerOptions
+        {
+            return services.AddOAuthHttpClient<HmacSha1RequestSigner, TOptions>(name, configureOptions);
         }
 
         /// <summary>
@@ -104,7 +170,7 @@ namespace GSS.Authorization.OAuth
         public static IHttpClientBuilder AddOAuthHttpClient(this IServiceCollection services, string name,
             Action<IServiceProvider, OAuthHttpHandlerOptions> configureOptions)
         {
-            return services.AddOAuthHttpClient<HmacSha1RequestSigner>(name, configureOptions);
+            return services.AddOAuthHttpClient<OAuthHttpHandlerOptions>(name, configureOptions);
         }
 
         private static void ValidateOptions(OAuthHttpHandlerOptions options)