+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2020-07-15
+
+### GSS.Authorization.OAuth.HttpClient 1.2.0
+
+- Support extended OAuthHttpHandlerOptions
+
 ## 2020-07-14
 
 ### GSS.Authorization.OAuth 1.1.0